fsm_work:
- Pulled ouit an FSM thread
- Moved FSM work to FSM thread
